Using 32-bit ints now for Int because grids can easily surpass the 16-bit limit

+++ src/Omega2D.h
@@ -10,7 +10,7 @@
 // Use this for indexes into panels or bodies
 // this means we can have no more than 65536 element segments in the system - seems reasonable for 2D
 #include <cstdint>
-using Int = uint16_t;
+using Int = uint32_t;
 #include <cstdlib>
 
 const size_t Dimensions = 2;
